requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
discord.py==2.3.2
python-dotenv==1.0.0
requests==2.31.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
aiohttp==3.9.1
